Tugas Akhir Semester 1/TrashBinBe/ai_model/requirements.txt
tensorflow>=2.10.0
pillow>=8.0.0
numpy>=1.21.0